+++ Utils.lua
@@ -11,3 +11,6 @@ local Config = ZenHUD.Config  -- Already exported by Config.lua
 -- Utilities
 --------------------------------------------------------------------------------
 local Utils = {}
+
+function Utils.Print(msg, debugOnly)
+    if debugOnly and not Config:Get("debug") then return end